{{ cookiecutter.repo_name }}/packages/azureml/pyproject.toml
[project]
name = "azureml"
version = "0.1.0"
requires-python = ">={{ cookiecutter.minimum_python_version }}"
description = "An azureml package for {{ cookiecutter.project_name }}" 
authors = [{ name = "{{ cookiecutter.author_name }}" }]
dependencies = [
  "src",
  "azure-ai-ml",
  "azure-identity",
]

[tool.uv.sources]
src = { workspace = true }

[tool.uv.workspace]
members = ["src/*"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
